+++ Barbuuuda.Models/Entities/Chat/DialogMemberEntity.cs
@@ -32,7 +32,7 @@ namespace Barbuuuda.Models.Entities.Chat
         /// <summary>
         /// Id диалога.
         /// </summary>
-        public int DialogId { get; set; }
+        public long DialogId { get; set; }
 
         [ForeignKey("Id")]
         public UserEntity User { get; set; }
